Draw position of individuals seen by a user on an image

Draw position of individuals seen by a user on an image

gg_users_image(x, image_id, filter_col, filter_val, path, img_name, buffer)

Arguments

x
spatial dataframe as issued from to_carto_segment

image_id
image_id

filter_col
optionnal. column on which to realize a filter

filter_val
optionnal. value for the filter.col. Can be a vector of values.

path
optionnal. Path where to save output image if set.

img_name
name used if path is not empty

buffer
optionnal. Numeric. Size of the buffer area to show.

Examples

# NOT RUN {
gg_users_image(x = ONC2_bucc_carto, image_id = 12937)
gg_users_image(x = ONC2_bucc_carto, filter_col = username, filter_val = "2olg", image_id = 12937)
# }